refactor: separate withCesium
