+++ src/core/withCesium.test.tsx
@@ -1,12 +1,12 @@
 import React from "react";
 import { mount } from "enzyme";
 
-import { withCesium } from "./context";
+import { withCesium } from "./withCesium";
 import { Provider } from "./context";
 
 describe("core/context", () => {
   it("should inject context to cesium prop", () => {
-    const Dummy: React.SFC<{ cesium: { dummy: string } }> = () => null;
+    const Dummy: React.FC<{ cesium: { dummy: string } }> = () => null;
     const WithCesiumDummy = withCesium<{}, { dummy: string }>(Dummy);
 
     const value = { dummy: "test" };
