+++ src/index.ts
@@ -3,6 +3,7 @@ export * from "./core/CameraOperation";
 export * from "./core/context";
 export * from "./core/events";
 export * from "./core/EventManager";
+export * from "./core/withCesium";
 
 export { default as Billboard } from "./Billboard";
 export { default as BillboardCollection } from "./BillboardCollection";
